Return the list of oceans and seas.

Return the list of oceans and seas.

accBasinNames(basin = NA)

Arguments

basin
A name of basin. Default is NA and returns a list of all the accepted names.

Value

A list of accepted names.

See also

Examples

accBasinNames()

#> $`Arctic Ocean`
#>  [1] "Arctic Ocean"              "Baffin Bay"               
#>  [3] "Barentsz Sea"              "Beaufort Sea"             
#>  [5] "Chukchi Sea"               "East Siberian Sea"        
#>  [7] "Greenland Sea"             "Hudson Bay"               
#>  [9] "Kara Sea"                  "Laptev Sea"               
#> [11] "Lincoln Sea"               "The Northwestern Passages"
#> [13] "White Sea"                
#> 
#> $`Atlantic Ocean`
#>  [1] "Adriatic Sea"                             
#>  [2] "Aegean Sea"                               
#>  [3] "Alboran Sea"                              
#>  [4] "Balearic (Iberian Sea)"                   
#>  [5] "Baltic Sea"                               
#>  [6] "Bay of Biscay"                            
#>  [7] "Bay of Fundy"                             
#>  [8] "Black Sea"                                
#>  [9] "Bristol Channel"                          
#> [10] "Caribbean Sea"                            
#> [11] "Celtic Sea"                               
#> [12] "Davis Strait"                             
#> [13] "English Channel"                          
#> [14] "Gulf of Bothnia"                          
#> [15] "Gulf of Finland"                          
#> [16] "Gulf of Guinea"                           
#> [17] "Gulf of Mexico"                           
#> [18] "Gulf of Riga"                             
#> [19] "Gulf of St. Lawrence"                     
#> [20] "Hudson Strait"                            
#> [21] "Inner Seas off the West Coast of Scotland"
#> [22] "Ionian Sea"                               
#> [23] "Irish Sea and St. George s Channel"       
#> [24] "Kattegat"                                 
#> [25] "Labrador Sea"                             
#> [26] "Ligurian Sea"                             
#> [27] "Mediterranean Sea - Eastern Basin"        
#> [28] "Mediterranean Sea - Western Basin"        
#> [29] "North Atlantic Ocean"                     
#> [30] "North Sea"                                
#> [31] "Norwegian Sea"                            
#> [32] "Rio de La Plata"                          
#> [33] "Sea of Azov"                              
#> [34] "Sea of Marmara"                           
#> [35] "Skagerrak"                                
#> [36] "South Atlantic Ocean"                     
#> [37] "Strait of Gibraltar"                      
#> [38] "Tyrrhenian Sea"                           
#> 
#> $`Indian Ocean`
#>  [1] "Andaman or Burma Sea"   "Arabian Sea"            "Bay of Bengal"         
#>  [4] "Great Australian Bight" "Gulf of Aden"           "Gulf of Aqaba"         
#>  [7] "Gulf of Oman"           "Gulf of Suez"           "Indian Ocean"          
#> [10] "Laccadive Sea"          "Malacca Strait"         "Mozambique Channel"    
#> [13] "Persian Gulf"           "Red Sea"                "Savu Sea"              
#> [16] "Timor Sea"             
#> 
#> $`Pacific Ocean`
#>  [1] "Arafura Sea"                                                
#>  [2] "Bali Sea"                                                   
#>  [3] "Banda Sea"                                                  
#>  [4] "Bass Strait"                                                
#>  [5] "Bering Sea"                                                 
#>  [6] "Bismarck Sea"                                               
#>  [7] "Celebes Sea"                                                
#>  [8] "Ceram Sea"                                                  
#>  [9] "Coral Sea"                                                  
#> [10] "Eastern China Sea"                                          
#> [11] "Flores Sea"                                                 
#> [12] "Gulf of Alaska"                                             
#> [13] "Gulf of Boni"                                               
#> [14] "Gulf of California"                                         
#> [15] "Gulf of Thailand"                                           
#> [16] "Gulf of Tomini"                                             
#> [17] "Halmahera Sea"                                              
#> [18] "Japan Sea"                                                  
#> [19] "Java Sea"                                                   
#> [20] "Makassar Strait"                                            
#> [21] "Molukka Sea"                                                
#> [22] "North Pacific Ocean"                                        
#> [23] "Philippine Sea"                                             
#> [24] "Sea of Okhotsk"                                             
#> [25] "Seto Naikai or Inland Sea"                                  
#> [26] "Singapore Strait"                                           
#> [27] "Solomon Sea"                                                
#> [28] "South China Sea"                                            
#> [29] "South Pacific Ocean"                                        
#> [30] "Sulu Sea"                                                   
#> [31] "Tasman Sea"                                                 
#> [32] "The Coastal Waters of Southeast Alaska and British Columbia"
#> [33] "Yellow Sea"                                                 
#> 
#> $`Southern Ocean`
#> [1] "Southern Ocean"
#> 

accBasinNames('Indian Ocean')

#> $`Indian Ocean`
#>  [1] "Andaman or Burma Sea"   "Arabian Sea"            "Bay of Bengal"         
#>  [4] "Great Australian Bight" "Gulf of Aden"           "Gulf of Aqaba"         
#>  [7] "Gulf of Oman"           "Gulf of Suez"           "Indian Ocean"          
#> [10] "Laccadive Sea"          "Malacca Strait"         "Mozambique Channel"    
#> [13] "Persian Gulf"           "Red Sea"                "Savu Sea"              
#> [16] "Timor Sea"             
#>