Describes all the variables available in the database.

Provides the index and the short and full names of all the variables available in the database.

accClimateVariables(v = NA)

Arguments

v
The name of a variable to quickly access its description and ID (default NA returns all possible values).

Value

A data frame descriptive of the climate variables available in the database (if v=NA) or the description of variable v.

Examples

accClimateVariables()

#>    Variable ID Variable name
#> 1            1          bio1
#> 2            2          bio2
#> 3            3          bio3
#> 4            4          bio4
#> 5            5          bio5
#> 6            6          bio6
#> 7            7          bio7
#> 8            8          bio8
#> 9            9          bio9
#> 10          10         bio10
#> 11          11         bio11
#> 12          12         bio12
#> 13          13         bio13
#> 14          14         bio14
#> 15          15         bio15
#> 16          16         bio16
#> 17          17         bio17
#> 18          18         bio18
#> 19          19         bio19
#> 20          20            ai
#>                                                  Description
#> 1                               Mean Annual Temperature (°C)
#> 2                                    Mean Diurnal Range (°C)
#> 3                                       Isothermality (x100)
#> 4     Temperature Seasonality (standard deviation x100) (°C)
#> 5                      Max Temperature of Warmest Month (°C)
#> 6                      Min Temperature of Coldest Month (°C)
#> 7                              Temperature Annual Range (°C)
#> 8                   Mean Temperature of Wettest Quarter (°C)
#> 9                    Mean Temperature of Driest Quarter (°C)
#> 10                  Mean Temperature of Warmest Quarter (°C)
#> 11                  Mean Temperature of Coldest Quarter (°C)
#> 12                                 Annual Precipitation (mm)
#> 13                       Precipitation of Wettest Month (mm)
#> 14                        Precipitation of Driest Month (mm)
#> 15 Precipitation Seasonality (Coefficient of Variation) (mm)
#> 16                     Precipitation of Wettest Quarter (mm)
#> 17                      Precipitation of Driest Quarter (mm)
#> 18                     Precipitation of Warmest Quarter (mm)
#> 19                     Precipitation of Coldest Quarter (mm)
#> 20                                             Aridity Index

accClimateVariables(v='bio12')

#> [1] "12"                        "bio12"                    
#> [3] "Annual Precipitation (mm)"