Return the list of the realms and associated biomes and ecoregions.

Return the list of the realms and associated biomes and ecoregions.

accRealmNames(realm = NA, ecoregion = TRUE)

Arguments

realm

A name of accepted realm. Default is NA and returns a list of all the biome and ecoregion names sorted by realm.

ecoregion

A boolean to choose whether to get the ecoregions names.

Value

A list with elements that correspond to the biomes (and possibly ecoregions) of each realm.

See also

Examples

accRealmNames(realm='Africotropical')
#> $Africotropical
#>                                                            biome
#> 1                                   Deserts and xeric shrublands
#> 2                                   Deserts and xeric shrublands
#> 3                                   Deserts and xeric shrublands
#> 4                                   Deserts and xeric shrublands
#> 5                                   Deserts and xeric shrublands
#> 6                                   Deserts and xeric shrublands
#> 7                                   Deserts and xeric shrublands
#> 8                                   Deserts and xeric shrublands
#> 9                                   Deserts and xeric shrublands
#> 10                                  Deserts and xeric shrublands
#> 11                                  Deserts and xeric shrublands
#> 12                                  Deserts and xeric shrublands
#> 13                                  Deserts and xeric shrublands
#> 14                                  Deserts and xeric shrublands
#> 15                                  Deserts and xeric shrublands
#> 16                                  Deserts and xeric shrublands
#> 17                                  Deserts and xeric shrublands
#> 18                                  Deserts and xeric shrublands
#> 19                                  Deserts and xeric shrublands
#> 20                                  Deserts and xeric shrublands
#> 21                                  Deserts and xeric shrublands
#> 22                               Flooded grasslands and savannas
#> 23                               Flooded grasslands and savannas
#> 24                               Flooded grasslands and savannas
#> 25                               Flooded grasslands and savannas
#> 26                               Flooded grasslands and savannas
#> 27                               Flooded grasslands and savannas
#> 28                               Flooded grasslands and savannas
#> 29                               Flooded grasslands and savannas
#> 30                                                         Lakes
#> 31                                                     Mangroves
#> 32                                                     Mangroves
#> 33                                                     Mangroves
#> 34                                                     Mangroves
#> 35                                                     Mangroves
#> 36                   Mediterranean Forests, woodlands and scrubs
#> 37                   Mediterranean Forests, woodlands and scrubs
#> 38                   Mediterranean Forests, woodlands and scrubs
#> 39                             Montane grasslands and shrublands
#> 40                             Montane grasslands and shrublands
#> 41                             Montane grasslands and shrublands
#> 42                             Montane grasslands and shrublands
#> 43                             Montane grasslands and shrublands
#> 44                             Montane grasslands and shrublands
#> 45                             Montane grasslands and shrublands
#> 46                             Montane grasslands and shrublands
#> 47                             Montane grasslands and shrublands
#> 48                             Montane grasslands and shrublands
#> 49                             Montane grasslands and shrublands
#> 50                             Montane grasslands and shrublands
#> 51                             Montane grasslands and shrublands
#> 52                             Montane grasslands and shrublands
#> 53                             Montane grasslands and shrublands
#> 54                 Temperate grasslands, savannas and shrublands
#> 55                 Temperate grasslands, savannas and shrublands
#> 56                 Temperate grasslands, savannas and shrublands
#> 57                Tropical and subtropical dry broadleaf forests
#> 58                Tropical and subtropical dry broadleaf forests
#> 59                Tropical and subtropical dry broadleaf forests
#> 60  Tropical and subtropical grasslands, savannas and shrublands
#> 61  Tropical and subtropical grasslands, savannas and shrublands
#> 62  Tropical and subtropical grasslands, savannas and shrublands
#> 63  Tropical and subtropical grasslands, savannas and shrublands
#> 64  Tropical and subtropical grasslands, savannas and shrublands
#> 65  Tropical and subtropical grasslands, savannas and shrublands
#> 66  Tropical and subtropical grasslands, savannas and shrublands
#> 67  Tropical and subtropical grasslands, savannas and shrublands
#> 68  Tropical and subtropical grasslands, savannas and shrublands
#> 69  Tropical and subtropical grasslands, savannas and shrublands
#> 70  Tropical and subtropical grasslands, savannas and shrublands
#> 71  Tropical and subtropical grasslands, savannas and shrublands
#> 72  Tropical and subtropical grasslands, savannas and shrublands
#> 73  Tropical and subtropical grasslands, savannas and shrublands
#> 74  Tropical and subtropical grasslands, savannas and shrublands
#> 75  Tropical and subtropical grasslands, savannas and shrublands
#> 76  Tropical and subtropical grasslands, savannas and shrublands
#> 77  Tropical and subtropical grasslands, savannas and shrublands
#> 78  Tropical and subtropical grasslands, savannas and shrublands
#> 79  Tropical and subtropical grasslands, savannas and shrublands
#> 80  Tropical and subtropical grasslands, savannas and shrublands
#> 81  Tropical and subtropical grasslands, savannas and shrublands
#> 82  Tropical and subtropical grasslands, savannas and shrublands
#> 83  Tropical and subtropical grasslands, savannas and shrublands
#> 84  Tropical and subtropical grasslands, savannas and shrublands
#> 85  Tropical and subtropical grasslands, savannas and shrublands
#> 86              Tropical and subtropical moist broadleaf forests
#> 87              Tropical and subtropical moist broadleaf forests
#> 88              Tropical and subtropical moist broadleaf forests
#> 89              Tropical and subtropical moist broadleaf forests
#> 90              Tropical and subtropical moist broadleaf forests
#> 91              Tropical and subtropical moist broadleaf forests
#> 92              Tropical and subtropical moist broadleaf forests
#> 93              Tropical and subtropical moist broadleaf forests
#> 94              Tropical and subtropical moist broadleaf forests
#> 95              Tropical and subtropical moist broadleaf forests
#> 96              Tropical and subtropical moist broadleaf forests
#> 97              Tropical and subtropical moist broadleaf forests
#> 98              Tropical and subtropical moist broadleaf forests
#> 99              Tropical and subtropical moist broadleaf forests
#> 100             Tropical and subtropical moist broadleaf forests
#> 101             Tropical and subtropical moist broadleaf forests
#> 102             Tropical and subtropical moist broadleaf forests
#> 103             Tropical and subtropical moist broadleaf forests
#> 104             Tropical and subtropical moist broadleaf forests
#> 105             Tropical and subtropical moist broadleaf forests
#> 106             Tropical and subtropical moist broadleaf forests
#> 107             Tropical and subtropical moist broadleaf forests
#> 108             Tropical and subtropical moist broadleaf forests
#> 109             Tropical and subtropical moist broadleaf forests
#> 110             Tropical and subtropical moist broadleaf forests
#> 111             Tropical and subtropical moist broadleaf forests
#> 112             Tropical and subtropical moist broadleaf forests
#> 113             Tropical and subtropical moist broadleaf forests
#> 114             Tropical and subtropical moist broadleaf forests
#> 115             Tropical and subtropical moist broadleaf forests
#>                                                 ecoregion
#> 1                              Aldabra Island xeric scrub
#> 2                    Arabian Peninsula coastal fog desert
#> 3                    East Saharan montane xeric woodlands
#> 4                                 Eritrean coastal desert
#> 5               Ethiopian xeric grasslands and shrublands
#> 6                     Gulf of Oman desert and semi-desert
#> 7                         Hobyo grasslands and shrublands
#> 8              Ile Europa and Bassas da India xeric scrub
#> 9                                  Kalahari xeric savanna
#> 10                                       Kaokoveld desert
#> 11                              Madagascar spiny thickets
#> 12                         Madagascar succulent woodlands
#> 13                  Masai xeric grasslands and shrublands
#> 14                                             Nama Karoo
#> 15                                           Namib desert
#> 16                             Namibian savanna woodlands
#> 17                        Socotra Island xeric shrublands
#> 18                         Somali montane xeric woodlands
#> 19                 Southwestern Arabian foothills savanna
#> 20                 Southwestern Arabian montane woodlands
#> 21                                        Succulent Karoo
#> 22                               East African halophytics
#> 23                                 Etosha Pan halophytics
#> 24                      Inner Niger Delta flooded savanna
#> 25                              Lake Chad flooded savanna
#> 26                             Saharan flooded grasslands
#> 27                      Zambezian coastal flooded savanna
#> 28                           Zambezian flooded grasslands
#> 29                                  Zambezian halophytics
#> 30                                                   Lake
#> 31                              Central African mangroves
#> 32                                 East African mangroves
#> 33                                      Guinean mangroves
#> 34                                   Madagascar mangroves
#> 35                              Southern Africa mangroves
#> 36                                        Albany thickets
#> 37                        Lowland fynbos and renosterveld
#> 38                        Montane fynbos and renosterveld
#> 39                Angolan montane forest-grassland mosaic
#> 40                    Angolan scarp savanna and woodlands
#> 41      Drakensberg alti-montane grasslands and woodlands
#> 42  Drakensberg montane grasslands, woodlands and forests
#> 43                         East African montane moorlands
#> 44       Eastern Zimbabwe montane forest-grassland mosaic
#> 45             Ethiopian montane grasslands and woodlands
#> 46                            Ethiopian montane moorlands
#> 47                                    Highveld grasslands
#> 48                    Jos Plateau forest-grassland mosaic
#> 49                            Madagascar ericoid thickets
#> 50             Maputaland-Pondoland bushland and thickets
#> 51                     Rwenzori-Virunga montane moorlands
#> 52          Southern Rift montane forest-grassland mosaic
#> 53           South Malawi montane forest-grassland mosaic
#> 54                             Al Hajar montane woodlands
#> 55  Amsterdam and Saint-Paul Islands temperate grasslands
#> 56    Tristan Da Cunha-Gough Islands shrub and grasslands
#> 57                         Cape Verde Islands dry forests
#> 58                       Madagascar dry deciduous forests
#> 59                    Zambezian Cryptosepalum dry forests
#> 60                               Angolan Miombo woodlands
#> 61                               Angolan Mopane woodlands
#> 62                         Ascension scrub and grasslands
#> 63                     Central Zambezian Miombo woodlands
#> 64                               Eastern Miombo woodlands
#> 65                                  East Sudanian savanna
#> 66                          Guinean forest-savanna mosaic
#> 67                                    Itigi-Sumbu thicket
#> 68                     Kalahari Acacia-Baikiaea woodlands
#> 69                                 Mandara Plateau mosaic
#> 70      Northern Acacia-Commiphora bushlands and thickets
#> 71               Northern Congolian forest-savanna mosaic
#> 72                                Sahelian Acacia savanna
#> 73                          Serengeti volcanic grasslands
#> 74        Somali Acacia-Commiphora bushlands and thickets
#> 75      Southern Acacia-Commiphora bushlands and thickets
#> 76                               Southern Africa bushveld
#> 77               Southern Congolian forest-savanna mosaic
#> 78                              Southern Miombo woodlands
#> 79                         St. Helena scrub and woodlands
#> 80                   Victoria Basin forest-savanna mosaic
#> 81                Western Congolian forest-savanna mosaic
#> 82                           Western Zambezian grasslands
#> 83                                  West Sudanian savanna
#> 84                         Zambezian and Mopane woodlands
#> 85                           Zambezian Baikiaea woodlands
#> 86                         Albertine Rift montane forests
#> 87                    Atlantic Equatorial coastal forests
#> 88                          Cameroonian Highlands forests
#> 89                      Central Congolian lowland forests
#> 90                                        Comoros forests
#> 91                         Cross-Niger transition forests
#> 92                     Cross-Sanaga-Bioko coastal forests
#> 93                           East African montane forests
#> 94                                    Eastern Arc forests
#> 95                        Eastern Congolian swamp forests
#> 96                                Eastern Guinean forests
#> 97                              Ethiopian montane forests
#> 98                            Granitic Seychelles forests
#> 99                                Guinean montane forests
#> 100                        Knysna-Amatole montane forests
#> 101                    KwaZulu-Cape coastal forest mosaic
#> 102                            Madagascar lowland forests
#> 103                           Madagascar subhumid forests
#> 104                      Maputaland coastal forest mosaic
#> 105                                     Mascarene forests
#> 106              Mount Cameroon and Bioko montane forests
#> 107                             Niger Delta swamp forests
#> 108                              Nigerian lowland forests
#> 109                Northeastern Congolian lowland forests
#> 110     Northern Zanzibar-Inhambane coastal forest mosaic
#> 111                Northwestern Congolian lowland forests
#> 112  Sao Tome, Principe and Annobon moist lowland forests
#> 113     Southern Zanzibar-Inhambane coastal forest mosaic
#> 114                       Western Congolian swamp forests
#> 115                       Western Guinean lowland forests
#> 
accRealmNames(realm='Africotropical', ecoregion=FALSE)
#> $Africotropical
#>  [1] "Deserts and xeric shrublands"                                
#>  [2] "Flooded grasslands and savannas"                             
#>  [3] "Lakes"                                                       
#>  [4] "Mangroves"                                                   
#>  [5] "Mediterranean Forests, woodlands and scrubs"                 
#>  [6] "Montane grasslands and shrublands"                           
#>  [7] "Temperate grasslands, savannas and shrublands"               
#>  [8] "Tropical and subtropical dry broadleaf forests"              
#>  [9] "Tropical and subtropical grasslands, savannas and shrublands"
#> [10] "Tropical and subtropical moist broadleaf forests"            
#>