Extract The climate values associated to a set of coordinates.

Extract The climate values associated to a set of coordinates.

climate_from_xy(
  long,
  lat,
  climate = accClimateVariables()[, 2],
  resol = 0.25,
  dbname = "gbif4crest_02"
)

Arguments

long
The longitude of the site.

lat
The latitude of the site.

climate
The climate variables to extract the values from. Returns all possible values by default.

resol
The resolution of the target climatology (default 0.25 degrees).

dbname
The name of the data source database.

Value

A data frame containing the climate values.

Examples

climate_from_xy(50, 10, c('bio1', 'ai'))

#>     bio1    ai
#> 1 25.985 0.021

climate_from_xy(50, 10)

#>     bio1  bio2   bio3    bio4   bio5   bio6   bio7   bio8   bio9  bio10 bio11
#> 1 25.985 9.627 61.686 234.019 33.585 17.979 15.606 26.621 28.377 28.435 23.03
#>    bio12  bio13 bio14  bio15  bio16 bio17  bio18 bio19    ai sst_ann sst_jfm
#> 1 63.199 15.131     0 90.367 31.266 1.236 11.139 5.418 0.021      NA      NA
#>   sst_amj sst_jas sst_ond sss_ann sss_jfm sss_amj sss_jas sss_ond icec_ann
#> 1      NA      NA      NA      NA      NA      NA      NA      NA       NA
#>   icec_jfm icec_amj icec_jas icec_ond diss_oxy nitrate phosphate silicate
#> 1       NA       NA       NA       NA       NA      NA        NA       NA