Extract distributions from the database

This function will extract the distributions of all the species composing each taxon and return them as a list.

crest.get_modern_data(
  pse,
  taxaType,
  climate,
  df = NA,
  ai.sqrt = FALSE,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  continents = NA,
  countries = NA,
  basins = NA,
  sectors = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  minGridCells = 20,
  elev_min = NA,
  elev_max = NA,
  elev_range = NA,
  year_min = 1900,
  year_max = 2021,
  nodate = TRUE,
  type_of_obs = c(1, 2, 3, 8, 9),
  selectedTaxa = NA,
  site_info = c(NA, NA),
  site_name = NA,
  dbname = "gbif4crest_02",
  verbose = TRUE
)

Arguments

pse
A pollen-Species equivalency table. See createPSE for details.

taxaType
A numerical index (between 1 and 6) to define the type of palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids, 4 for foraminifers, 5 for diatoms and 6 for rodents. The example dataset uses taxaType=0 (pseudo-data). Default is 1.

climate
A vector of the climate variables to extract. See accClimateVariables for the list of accepted values.

df
A data frame containing the data to reconstruct (counts, percentages or presence/absence data).

ai.sqrt
A boolean to indicate whether ai values should be square-root transformed (default FALSE).

xmn
The coordinates defining the study area.

xmx
The coordinates defining the study area.

ymn
The coordinates defining the study area.

ymx
The coordinates defining the study area.

continents
A vector of the continent names defining the study area.

countries
A vector of the country names defining the study area.

basins
A vector of the ocean names defining the study area.

sectors
A vector of the marine sector names defining the study area.

realms
A vector of the studied botanical realms defining the study area.

biomes
A vector of the studied botanical biomes defining the study area.

ecoregions
A vector of the studied botanical ecoregions defining the study area.

minGridCells
The minimum number of unique presence data necessary to estimate a species' climate response. Default is 20.

elev_min
Parameters to only selected grid cells with an elevation higher than elev_min or lower than elev_max (default is 'NA ).

elev_max
Parameters to only selected grid cells with an elevation higher than elev_min or lower than elev_max (default is 'NA ).

elev_range
Parameters discard the grid cell with a high elevation range (default is NA).

year_min
The oldest and youngest occurrences accepted (default is 1900-2021).

year_max
The oldest and youngest occurrences accepted (default is 1900-2021).

nodate
A boolean to accept occurrences without a date (can overlap with occurrences with a date; default TRUE).

type_of_obs
The type of observation to use in the study. 1: human observation, 2: observation, 3: preserved specimen, 4: living specimen, 5: fossil specimen, 6: material sample, 7: machine observation, 8: literature, 9: unknown (Default c(1, 2, 3, 8, 9))

selectedTaxa
A data frame assigns which taxa should be used for each variable (1 if the taxon should be used, 0 otherwise). The colnames should be the climate variables' names and the rownames the taxa names. Default is 1 for all taxa and all variables.

site_info
A vector containing the coordinates of the study site. Default c(NA, NA).

site_name
The name of the dataset (default NA).

dbname
The name of the database. Default is 'gbif4crest_02'.

verbose
A boolean to print non-essential comments on the terminal (default TRUE).

Value

A crestObj object containing the spatial distributions.

Examples

data(crest_ex_pse)
data(crest_ex_selection)
data(crest_ex)
x <- crest.get_modern_data( df = crest_ex,
  pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example",
  verbose = FALSE
)

#> Warning: The classification of one or more taxa into species was not successful. Check `x$misc$taxa_notes` for details.

x

#> *
#> * Summary of the crestObj named ``:
#> *   x Calibration data formatted .. TRUE
#> *   x PDFs fitted ................. FALSE
#> *   x Climate reconstructed ....... FALSE
#> *   x Leave-One-Out analysis ...... FALSE
#> *
#> * The dataset to be reconstructed (`df`) is composed of 20 samples with 6 taxa.
#> * Variables to analyse: bio1, bio12
#> *
#> * The calibration dataset was defined using the following set of parameters:
#> *   x Proxy type ............ Example dataset
#> *   x Longitude ............. [0 - 15]
#> *   x Latitude .............. [0 - 15]
#> *
#> * Of the 7 taxa provided in `df` and `PSE`, 1 cannot be analysed.
#> * (This may be expected, but check `$misc$taxa_notes` for additional details.)
#> *

lapply(x$modelling$distributions, head)

#> $Taxon1
#>   taxonid longitude latitude bio1 bio12
#> 1       1     12.75     1.75 23.8   277
#> 2       1     11.75     2.75 23.7   172
#> 3       1     13.75     2.75 24.0    93
#> 4       1     14.25     3.75 25.6   203
#> 5       1     14.75     4.25 26.4   171
#> 6       1     11.25     4.75 22.5   281
#> 
#> $Taxon2
#>   taxonid longitude latitude bio1 bio12
#> 1       2      7.25     0.25 21.0     5
#> 2       2     13.25     0.25 26.9    54
#> 3       2     14.75     0.25 23.7    97
#> 4       2     13.75     1.75 26.6   220
#> 5       2     14.25     2.25 24.1   149
#> 6       2     11.25     2.75 22.4   186
#> 
#> $Taxon3
#>   taxonid longitude latitude bio1 bio12
#> 1       3      0.25     0.25 15.4     3
#> 2       3      0.75     0.25 13.6    30
#> 3       3      3.75     0.25 16.5    84
#> 4       3      7.25     0.25 21.0     5
#> 5       3      1.25     0.75 16.0    52
#> 6       3      2.75     0.75 16.1   190
#> 
#> $Taxon4
#>   taxonid longitude latitude bio1 bio12
#> 1       4     13.25     0.25 26.9    54
#> 2       4      9.75     0.75 23.0   158
#> 3       4     12.25     1.25 22.1   153
#> 4       4      8.25     3.75 21.6   257
#> 5       4      9.75     3.75 19.4   201
#> 6       4     10.25     3.75 22.5   356
#> 
#> $Taxon5
#>   taxonid longitude latitude bio1 bio12
#> 1       5      1.25     0.25 16.8    83
#> 2       5      1.75     0.25 14.9     4
#> 3       5      2.25     0.25 17.6    97
#> 4       5      2.75     0.25 19.0    12
#> 5       5      3.25     0.25 18.1    57
#> 6       5      3.75     0.25 16.5    84
#> 
#> $Taxon6
#>   taxonid longitude latitude bio1 bio12
#> 1       6      0.25     0.25 15.4     3
#> 2       6      0.75     0.25 13.6    30
#> 3       6      1.25     0.25 16.8    83
#> 4       6      1.75     0.25 14.9     4
#> 5       6      2.25     0.25 17.6    97
#> 6       6      2.75     0.25 19.0    12
#>