Create a subset of the global calibration dataset

Create a subset of the global calibration dataset

dbSubset(
  taxaType,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  dbname = "gbif4crest_02",
  out = "gbif4crest_reduced",
  verbose = TRUE
)

Arguments

taxaType

A numerical index (between 1 and 6) to define the type of palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids, 4 for foraminifers, 5 for diatoms and 6 for rodents. The example dataset uses taxaType=0 (pseudo-data). Default is 1.

xmn, xmx, ymn, ymx

The coordinates defining the study area.

dbname

The name of the database. Default is 'gbif4crest_02' and data will be extracted from the online database. The SQLite3 version of the database can also be used here by providing the complete path to a file ending by .sqlite3, e.g. /path/to/file/gbif4crest_02.sqlite3

out

The name or path of the new dataset

verbose

A boolean to print non-essential comments on the terminal (default TRUE).

See also

Examples

if (FALSE) {
  dbSubset(2, xmn=0, xmx=15, ymn=0, ymx=15, out='example.sqlite3')
}