Extract the distribution of the studied climate gradient(s) across the study area.

Extract the distribution of the studied climate gradient(s) across the study area.

getClimateSpace(
  climate,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  continents = NA,
  countries = NA,
  basins = NA,
  sectors = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  elev_min = NA,
  elev_max = NA,
  elev_range = NA,
  dbname = "gbif4crest_02"
)

Arguments

climate

A vector of the climate variables to extract. See accClimateVariables for the list of accepted values.

xmn, xmx, ymn, ymx

The coordinates defining the study area.

continents

A vector of the continent names defining the study area.

countries

A vector of the country names defining the study area.

basins

A vector of the ocean names defining the study area.

sectors

A vector of the marine sector names defining the study area.

realms

A vector of the studied botanical realms defining the study area.

biomes

A vector of the studied botanical biomes defining the study area.

ecoregions

A vector of the studied botanical ecoregions defining the study area.

elev_min, elev_max

Parameters to only selected grid cells with an elevation higher than elev_min or lower than elev_max (default is 'NA ).

elev_range

Parameters discard the grid cell with a high elevation range (default is NA).

dbname

The name of the data source database.

Value

A matrix of occurrence records with the associated climate.

See also

Examples

if (FALSE) {
  climate <- getClimateSpace("bio1", -90, 90, -90, 90,
    continents = "Europe",
    countries = c("Germany", "Netherlands", "Sweden"),
    realms = "Palaearctic"
  )
  head(climate)
  terra::plot(terra::rast(climate, type='xyz'), asp=1)
}