Extract taxonID(s) corresponding to the taxonomic description

Extract all possible TaxonIDs corresponding to the provided taxonomical description, which can be at the family, the genus or the species levels.

getTaxonomy(
  family = "",
  genus = "",
  species = "",
  taxaType = 1,
  depth.out = 8,
  dbname = "gbif4crest_02"
)

Arguments

family
The name of the family.

genus
The name of the genus.

species
The name of the species.

taxaType
A numerical index (between 1 and 5) to define the type of palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids, 4 for foraminifers, 5 for diatoms and 6 for rodents.

depth.out
The taxonomic resolution of the output table. 1 for Kingdom, 2 for phylum, 3 for class_name, 4 for order_name, 5 for family, 6 for genus, 7 for species and 8 to also include the taxonID.

dbname
The name of the database. Default is 'gbif4crest_02'.

Value

A vector of unique taxonIDs.

Examples

getTaxonomy("Zamiaceae")

#>     kingdom       phylum  class_name order_name    family           genus
#> 1   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bororoa
#> 2   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bororoa
#> 3   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bowenia
#> 4   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bowenia
#> 5   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bowenia
#> 6   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae         Bowenia
#> 7   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae          Brunoa
#> 8   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 9   Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 10  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 11  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 12  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 13  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 14  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 15  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 16  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 17  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 18  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 19  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 20  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 21  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 22  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 23  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 24  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 25  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 26  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 27  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 28  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 29  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 30  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 31  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 32  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 33  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 34  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 35  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 36  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Ceratozamia
#> 37  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 38  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 39  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 40  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 41  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 42  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 43  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 44  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 45  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 46  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 47  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 48  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 49  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 50  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 51  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Dioon
#> 52  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 53  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 54  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 55  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 56  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 57  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 58  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 59  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 60  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 61  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 62  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 63  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 64  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 65  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 66  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 67  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 68  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 69  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 70  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 71  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 72  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 73  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 74  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 75  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 76  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 77  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 78  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 79  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 80  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 81  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 82  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 83  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 84  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 85  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 86  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 87  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 88  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 89  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 90  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 91  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 92  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 93  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 94  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 95  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 96  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 97  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 98  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 99  Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 100 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 101 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 102 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 103 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 104 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 105 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 106 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 107 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 108 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 109 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 110 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 111 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 112 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 113 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 114 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 115 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae   Encephalartos
#> 116 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Lepidozamia
#> 117 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Lepidozamia
#> 118 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae     Lepidozamia
#> 119 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 120 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 121 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 122 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 123 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 124 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 125 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 126 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 127 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 128 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 129 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 130 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 131 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 132 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 133 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 134 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 135 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 136 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 137 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 138 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 139 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 140 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 141 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 142 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 143 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 144 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 145 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 146 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 147 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 148 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 149 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 150 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 151 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 152 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 153 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 154 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 155 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 156 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 157 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 158 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 159 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 160 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Macrozamia
#> 161 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Microcycas
#> 162 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Pterostoma
#> 163 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Pterostoma
#> 164 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Pterostoma
#> 165 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Pterostoma
#> 166 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae Restrepophyllum
#> 167 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae       Stangeria
#> 168 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae      Worsdellia
#> 169 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 170 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 171 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 172 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 173 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 174 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 175 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 176 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 177 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 178 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 179 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 180 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 181 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 182 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 183 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 184 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 185 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 186 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 187 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 188 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 189 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 190 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 191 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 192 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 193 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 194 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 195 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 196 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 197 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 198 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 199 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 200 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 201 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 202 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 203 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 204 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 205 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 206 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 207 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 208 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 209 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 210 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 211 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 212 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 213 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 214 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 215 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 216 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 217 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 218 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 219 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 220 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 221 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 222 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 223 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 224 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 225 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 226 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 227 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 228 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 229 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 230 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 231 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 232 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 233 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 234 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 235 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 236 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 237 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 238 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 239 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 240 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 241 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#> 242 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae           Zamia
#>                               species taxonid
#> 1                   Bororoa andreisii 1000358
#> 2                Bororoa anzulovichii 1000359
#> 3                    Bowenia eocenica 1000360
#> 4                   Bowenia papillosa 1000365
#> 5                   Bowenia serrulata 1000101
#> 6                 Bowenia spectabilis 1000043
#> 7               Brunoa santarrosensis 1000374
#> 8               Ceratozamia alvarezii 1000172
#> 9              Ceratozamia brevifrons 1000184
#> 10          Ceratozamia chimalapensis 1000210
#> 11              Ceratozamia decumbens 1000198
#> 12          Ceratozamia euryphyllidia 1000201
#> 13        Ceratozamia floersheimensis 1000356
#> 14           Ceratozamia fuscoviridis 1000343
#> 15                 Ceratozamia hildae 1000054
#> 16            Ceratozamia hondurensis 1000334
#> 17            Ceratozamia huastecorum 1000199
#> 18            Ceratozamia kuesteriana 1000192
#> 19              Ceratozamia latifolia 1000186
#> 20                Ceratozamia matudae 1000167
#> 21               Ceratozamia mexicana 1000183
#> 22          Ceratozamia microstrobila 1000185
#> 23             Ceratozamia miqueliana 1000187
#> 24               Ceratozamia mirandae 1000158
#> 25               Ceratozamia mixeorum 1000148
#> 26               Ceratozamia morettii 1000196
#> 27              Ceratozamia norstogii 1000157
#> 28                Ceratozamia robusta 1000057
#> 29                Ceratozamia sabatoi 1000191
#> 30            Ceratozamia santillanii 1000175
#> 31         Ceratozamia subroseophylla 1000077
#> 32                 Ceratozamia tenuis 1000208
#> 33              Ceratozamia vovidesii 1000147
#> 34          Ceratozamia whitelockiana 1000194
#> 35              Ceratozamia zaragozae 1000178
#> 36               Ceratozamia zoquorum 1000171
#> 37                Dioon angustifolium 1000052
#> 38                    Dioon argenteum 1000205
#> 39                    Dioon califanoi 1000181
#> 40                      Dioon caputoi 1000206
#> 41                        Dioon edule 1000060
#> 42                   Dioon holmgrenii 1000204
#> 43                       Dioon mejiae 1000162
#> 44                      Dioon merolae 1000170
#> 45                  Dioon planifolium 1000146
#> 46                     Dioon purpusii 1000203
#> 47                   Dioon rzedowskii 1000182
#> 48                    Dioon sonorense 1000188
#> 49                   Dioon spinulosum 1000189
#> 50                  Dioon stevensonii 1000143
#> 51                   Dioon tomasellii 1000142
#> 52             Encephalartos aemulans 1000307
#> 53         Encephalartos altensteinii 1000051
#> 54            Encephalartos aplanatus 1000271
#> 55            Encephalartos arenarius 1000277
#> 56              Encephalartos barteri 1000239
#> 57      Encephalartos brevifoliolatus 1000298
#> 58            Encephalartos bubalinus 1000322
#> 59               Encephalartos caffer 1000269
#> 60              Encephalartos cerinus 1000274
#> 61     Encephalartos chimanimaniensis 1000312
#> 62            Encephalartos concinnus 1000316
#> 63              Encephalartos cupidus 1000297
#> 64         Encephalartos cycadifolius 1000202
#> 65            Encephalartos delucanus 1000351
#> 66         Encephalartos denticulatus 1000425
#> 67          Encephalartos dolomiticus 1000300
#> 68            Encephalartos dyerianus 1000296
#> 69         Encephalartos equatorialis 1000295
#> 70      Encephalartos eugene-maraisii 1000302
#> 71                Encephalartos ferox 1000152
#> 72  Encephalartos friderici-guilielmi 1000270
#> 73          Encephalartos ghellinckii 1000218
#> 74               Encephalartos gratus 1000240
#> 75             Encephalartos heenanii 1000305
#> 76        Encephalartos hildebrandtii 1000066
#> 77             Encephalartos hirsutus 1000323
#> 78             Encephalartos horridus 1000220
#> 79              Encephalartos humilis 1000320
#> 80             Encephalartos inopinus 1000265
#> 81           Encephalartos ituriensis 1000273
#> 82              Encephalartos kisambo 1000310
#> 83          Encephalartos laevifolius 1000303
#> 84              Encephalartos lanatus 1000309
#> 85            Encephalartos latifrons 1000293
#> 86         Encephalartos laurentianus 1000268
#> 87         Encephalartos lebomboensis 1000221
#> 88            Encephalartos lehmannii 1000215
#> 89          Encephalartos longifolius 1000214
#> 90       Encephalartos macrostrobilus 1000304
#> 91           Encephalartos manikensis 1000151
#> 92         Encephalartos marunguensis 1000313
#> 93      Encephalartos middelburgensis 1000301
#> 94            Encephalartos msinganus 1000306
#> 95              Encephalartos munchii 1000311
#> 96           Encephalartos natalensis 1000166
#> 97             Encephalartos ngoyanus 1000324
#> 98         Encephalartos nubimontanus 1000278
#> 99        Encephalartos paucidentatus 1000266
#> 100              Encephalartos poggei 1000222
#> 101            Encephalartos princeps 1000279
#> 102          Encephalartos pterogonus 1000318
#> 103            Encephalartos relictus 1000314
#> 104           Encephalartos schmitzii 1000319
#> 105             Encephalartos sclavoi 1000275
#> 106          Encephalartos senticosus 1000232
#> 107     Encephalartos septentrionalis 1000338
#> 108          Encephalartos tegulaneus 1000160
#> 109        Encephalartos transvenosus 1000219
#> 110         Encephalartos trispinosus 1000154
#> 111             Encephalartos turneri 1000299
#> 112       Encephalartos umbeluziensis 1000233
#> 113            Encephalartos villosus 1000040
#> 114         Encephalartos whitelockii 1000308
#> 115              Encephalartos woodii 1000317
#> 116             Lepidozamia foveolata 1000366
#> 117                 Lepidozamia hopei 1000088
#> 118          Lepidozamia peroffskyana 1000049
#> 119           Macrozamia cardiacensis 1000115
#> 120               Macrozamia communis 1000041
#> 121               Macrozamia concinna 1000106
#> 122               Macrozamia conferta 1000103
#> 123                 Macrozamia cranei 1000099
#> 124            Macrozamia crassifolia 1000112
#> 125              Macrozamia diplomera 1000126
#> 126              Macrozamia douglasii 1000064
#> 127                  Macrozamia dyeri 1000069
#> 128                Macrozamia elegans 1000095
#> 129              Macrozamia fawcettii 1000065
#> 130             Macrozamia fearnsidei 1000114
#> 131               Macrozamia flexuosa 1000047
#> 132                Macrozamia fraseri 1000063
#> 133           Macrozamia glaucophylla 1000111
#> 134             Macrozamia heteromera 1000100
#> 135                Macrozamia humilis 1000098
#> 136              Macrozamia johnsonii 1000104
#> 137           Macrozamia lomandroides 1000075
#> 138             Macrozamia longispina 1000130
#> 139                 Macrozamia lucida 1000059
#> 140           Macrozamia macdonnellii 1000081
#> 141               Macrozamia machinii 1000118
#> 142               Macrozamia macleayi 1000083
#> 143               Macrozamia miquelii 1000097
#> 144                Macrozamia montana 1000132
#> 145                 Macrozamia moorei 1000105
#> 146        Macrozamia mountperriensis 1000079
#> 147                Macrozamia occidua 1000102
#> 148             Macrozamia parcifolia 1000108
#> 149        Macrozamia pauli-guilielmi 1000039
#> 150           Macrozamia platyrhachis 1000123
#> 151            Macrozamia plurinervia 1000087
#> 152             Macrozamia polymorpha 1000091
#> 153                Macrozamia reducta 1000109
#> 154                Macrozamia reidlei 1000280
#> 155                Macrozamia riedlei 1000067
#> 156                Macrozamia secunda 1000089
#> 157             Macrozamia serpentina 1000125
#> 158               Macrozamia spiralis 1000045
#> 159              Macrozamia stenomera 1000090
#> 160                Macrozamia viridis 1000128
#> 161               Microcycas calocoma 1000230
#> 162           Pterostoma anastomosans 1000363
#> 163              Pterostoma douglasii 1000362
#> 164               Pterostoma hirsutus 1000361
#> 165              Pterostoma zamioides 1000364
#> 166        Restrepophyllum chiguoides 1000370
#> 167                 Stangeria eriopus 1000078
#> 168              Worsdellia bonettiae 1000373
#> 169                   Zamia acuminata 1000350
#> 170            Zamia allison-aumourii 1000391
#> 171                    Zamia amazonum 1000226
#> 172                  Zamia amplifolia 1000344
#> 173                Zamia angustifolia 1000276
#> 174                   Zamia boliviana 1000267
#> 175                      Zamia chigua 1000207
#> 176                 Zamia cremnophila 1000177
#> 177                     Zamia cunaria 1000256
#> 178                   Zamia decumbens 1000197
#> 179                     Zamia disodon 1000258
#> 180                   Zamia dressleri 1000241
#> 181               Zamia elegantissima 1000212
#> 182            Zamia encephalartoides 1000242
#> 183                       Zamia erosa 1000080
#> 184               Zamia fairchildiana 1000042
#> 185                    Zamia fischeri 1000180
#> 186                  Zamia furfuracea 1000050
#> 187                     Zamia gentryi 1000345
#> 188                   Zamia gomeziana 1000387
#> 189                    Zamia herrerae 1000174
#> 190                   Zamia huilensis 1000236
#> 191             Zamia hymenophyllidia 1000261
#> 192                  Zamia imperialis 1000294
#> 193                   Zamia incognita 1000335
#> 194                     Zamia inermis 1000193
#> 195                Zamia integrifolia 1000038
#> 196                  Zamia ipetiensis 1000347
#> 197                  Zamia katzeriana 1000190
#> 198                   Zamia lacandona 1000155
#> 199                   Zamia lecointei 1000257
#> 200                    Zamia lindenii 1000149
#> 201                    Zamia lindleyi 1000386
#> 202                  Zamia loddigesii 1000163
#> 203                    Zamia lucayana 1000348
#> 204                    Zamia manicata 1000225
#> 205                   Zamia meermanii 1000211
#> 206              Zamia melanorrhachis 1000224
#> 207                     Zamia montana 1000237
#> 208                   Zamia monticola 1000213
#> 209                    Zamia muricata 1000251
#> 210                        Zamia nana 1000325
#> 211                   Zamia nesophila 1000164
#> 212              Zamia neurophyllidia 1000161
#> 213                     Zamia obliqua 1000223
#> 214                  Zamia oligodonta 1000238
#> 215                    Zamia oreillyi 1000388
#> 216                   Zamia paucijuga 1000076
#> 217                 Zamia poeppigiana 1000339
#> 218               Zamia portoricensis 1000061
#> 219                     Zamia prasina 1000144
#> 220             Zamia pseudomonticola 1000247
#> 221            Zamia pseudoparasitica 1000216
#> 222                      Zamia pumila 1000056
#> 223                    Zamia purpurea 1000173
#> 224                     Zamia pygmaea 1000200
#> 225                  Zamia pyrophylla 1000165
#> 226                   Zamia restrepoi 1000243
#> 227                     Zamia roezlii 1000086
#> 228                  Zamia sandovalii 1000331
#> 229                    Zamia skinneri 1000209
#> 230               Zamia soconuscensis 1000168
#> 231                     Zamia spartea 1000176
#> 232                   Zamia splendens 1000169
#> 233                   Zamia standleyi 1000058
#> 234                  Zamia tolimensis 1000333
#> 235                Zamia tuerckheimii 1000346
#> 236                        Zamia ulei 1000150
#> 237                        Zamia urep 1000244
#> 238                   Zamia variegata 1000195
#> 239                   Zamia vasquezii 1000272
#> 240                   Zamia vazquezii 1000179
#> 241              Zamia verschaffeltii 1000349
#> 242                    Zamia wallisii 1000341

getTaxonomy(genus="Ceratozamia", depth.out=6)

#>   kingdom       phylum  class_name order_name    family       genus
#> 1 Plantae Tracheophyta Cycadopsida  Cycadales Zamiaceae Ceratozamia

## \code{taxaType = 2} searches for beetles and not plants, so the next line returns nothing.
getTaxonomy("Zamiaceae", "Ceratozamia", taxaType = 2) 

#> [1] kingdom    phylum     class_name order_name family     genus      species   
#> [8] taxonid   
#> <0 rows> (or 0-length row.names)