Calculate the climate tolerance of the taxa from their pdfs.

Calculate the climate tolerance of the taxa from their pdfs.

pdf_ranges(
  x,
  climate = x$parameters$climate,
  taxanames = x$input$taxa.name,
  uncertainties = x$parameters$uncertainties,
  orderby = NA
)

Arguments

x

A crestObj generated by either the crest.calibrate, crest.reconstruct or crest functions.

climate

Climate variables to be used to generate the plot. By default all the variables are included.

taxanames

A list of taxa to use for the plot (default is all the recorded taxa).

uncertainties

A (vector of) threshold value(s) indicating the error bars that should be calculated (default are the values stored in x).

orderby

A string ('name', or one of the climate variables) to sort the output table

Value

The set of coordinates ext projected in equal earth.

Examples

data(reconstr)
pdf_ranges(reconstr, climate='bio1')
#> $`Range = 50%`
#>        bio1_tol_inf bio1_tol_sup bio1_range
#> Taxon1     23.00601     25.33066   2.324649
#> Taxon2     23.64729     25.81162   2.164329
#> Taxon3     14.42886     16.99399   2.565130
#> Taxon4     18.35671     22.20441   3.847695
#> Taxon5     18.11623     22.52505   4.408818
#> Taxon6     18.11623     22.36473   4.248497
#> 
#> $`Range = 95%`
#>        bio1_tol_inf bio1_tol_sup bio1_range
#> Taxon1     20.68136     27.73547   7.054108
#> Taxon2     21.56313     27.97595   6.412826
#> Taxon3     11.86373     19.55912   7.695391
#> Taxon4     14.50902     25.97194  11.462926
#> Taxon5     13.86774     26.85371  12.985972
#> Taxon6     13.86774     26.53307  12.665331
#> 
pdf_ranges(reconstr, climate='bio12', orderby='bio1', uncertainties=c(0.2, 0.6, 0.95))
#> Warning: The value orderby = 'bio1' is not accepted. It should be 'name' or one of the climate variables. The data were not sorted.
#> $`Range = 20%`
#>        bio12_tol_inf bio12_tol_sup bio12_range
#> Taxon1     620.04008      687.7756    67.73547
#> Taxon2     312.62525      414.2285   101.60321
#> Taxon3     312.62525      416.8337   104.20842
#> Taxon4     719.03808      781.5631    62.52505
#> Taxon5      80.76152      119.8397    39.07816
#> Taxon6     247.49499      336.0721    88.57715
#> 
#> $`Range = 60%`
#>        bio12_tol_inf bio12_tol_sup bio12_range
#> Taxon1      544.4890      781.5631    237.0741
#> Taxon2      224.0481      583.5671    359.5190
#> Taxon3      224.0481      586.1723    362.1242
#> Taxon4      651.3026      864.9299    213.6273
#> Taxon5       49.4990      192.7856    143.2866
#> Taxon6      171.9439      487.1743    315.2305
#> 
#> $`Range = 95%`
#>        bio12_tol_inf bio12_tol_sup bio12_range
#> Taxon1     429.85972      995.1904    565.3307
#> Taxon2     127.65531     1021.2425    893.5872
#> Taxon3     127.65531     1029.0581    901.4028
#> Taxon4     536.67335     1047.2946    510.6212
#> Taxon5      20.84168      458.5170    437.6754
#> Taxon6      91.18236      919.6393    828.4569
#>