Return the list of the continents and associated countries.

Return the list of the continents and associated countries.

accCountryNames(continent = NA)

Arguments

continent
A name of continent. Default is NA and returns a list of all the country names sorted by continent.

Value

A list where each element is a vector of corresponding country names.

See also

Examples

accCountryNames()

#> $Africa
#>  [1] "Algeria"                  "Angola"                  
#>  [3] "Benin"                    "Botswana"                
#>  [5] "Burkina Faso"             "Burundi"                 
#>  [7] "Cabo Verde"               "Cameroon"                
#>  [9] "Central African Republic" "Chad"                    
#> [11] "Comoros"                  "Congo"                   
#> [13] "Côte d Ivoire"            "Djibouti"                
#> [15] "DRC"                      "Egypt"                   
#> [17] "Equatorial Guinea"        "Eritrea"                 
#> [19] "eSwatini"                 "Ethiopia"                
#> [21] "Gabon"                    "Gambia"                  
#> [23] "Ghana"                    "Guinea"                  
#> [25] "Guinea-Bissau"            "Kenya"                   
#> [27] "Lesotho"                  "Liberia"                 
#> [29] "Libya"                    "Madagascar"              
#> [31] "Malawi"                   "Mali"                    
#> [33] "Mauritania"               "Mauritius"               
#> [35] "Mayotte"                  "Morocco"                 
#> [37] "Mozambique"               "Namibia"                 
#> [39] "Niger"                    "Nigeria"                 
#> [41] "Réunion"                  "Rwanda"                  
#> [43] "Saint Helena"             "Sao Tome and Principe"   
#> [45] "Senegal"                  "Seychelles"              
#> [47] "Sierra Leone"             "Somalia"                 
#> [49] "South Africa"             "South Sudan"             
#> [51] "Sudan"                    "Tanzania"                
#> [53] "Togo"                     "Tunisia"                 
#> [55] "Uganda"                   "Western Sahara"          
#> [57] "Zambia"                   "Zimbabwe"                
#> 
#> $Antarctica
#> [1] "Antarctica"                        "Bouvet Island"                    
#> [3] "French Southern Territories"       "Heard Island and McDonald Islands"
#> 
#> $Asia
#>  [1] "Afghanistan"                    "Armenia"                       
#>  [3] "Azerbaijan"                     "Bahrain"                       
#>  [5] "Bangladesh"                     "Bhutan"                        
#>  [7] "British Indian Ocean Territory" "Brunei Darussalam"             
#>  [9] "Cambodia"                       "China"                         
#> [11] "Christmas Island"               "Cocos (Keeling) Islands"       
#> [13] "Cyprus"                         "Georgia"                       
#> [15] "Hong Kong"                      "India"                         
#> [17] "Indonesia"                      "Iran"                          
#> [19] "Iraq"                           "Israel"                        
#> [21] "Japan"                          "Jordan"                        
#> [23] "Kazakhstan"                     "Kuwait"                        
#> [25] "Kyrgyzstan"                     "Laos"                          
#> [27] "Lebanon"                        "Macao"                         
#> [29] "Malaysia"                       "Maldives"                      
#> [31] "Mongolia"                       "Myanmar"                       
#> [33] "Nepal"                          "North Korea"                   
#> [35] "Oman"                           "Pakistan"                      
#> [37] "Palestine"                      "Philippines"                   
#> [39] "Qatar"                          "Saudi Arabia"                  
#> [41] "Singapore"                      "South Korea"                   
#> [43] "Spratly Islands"                "Sri Lanka"                     
#> [45] "Syria"                          "Taiwan"                        
#> [47] "Tajikistan"                     "Thailand"                      
#> [49] "Timor-Leste"                    "Turkey"                        
#> [51] "Turkmenistan"                   "United Arab Emirates"          
#> [53] "Uzbekistan"                     "Vietnam"                       
#> [55] "Yemen"                         
#> 
#> $Europe
#>  [1] "Åland Islands"                 "Albania"                      
#>  [3] "Andorra"                       "Austria"                      
#>  [5] "Belarus"                       "Belgium"                      
#>  [7] "Bosnia and Herzegovina"        "Bulgaria"                     
#>  [9] "Croatia"                       "Czechia"                      
#> [11] "Denmark"                       "Estonia"                      
#> [13] "Faroe Islands"                 "Finland"                      
#> [15] "France"                        "Germany"                      
#> [17] "Gibraltar"                     "Greece"                       
#> [19] "Guernsey"                      "Holy See (Vatican City State)"
#> [21] "Hungary"                       "Iceland"                      
#> [23] "Ireland"                       "Isle of Man"                  
#> [25] "Italy"                         "Jersey"                       
#> [27] "Kosovo"                        "Latvia"                       
#> [29] "Liechtenstein"                 "Lithuania"                    
#> [31] "Luxembourg"                    "Macedonia"                    
#> [33] "Malta"                         "Moldova"                      
#> [35] "Monaco"                        "Montenegro"                   
#> [37] "Netherlands"                   "Norway"                       
#> [39] "Poland"                        "Portugal"                     
#> [41] "Romania"                       "Russia"                       
#> [43] "San Marino"                    "Serbia"                       
#> [45] "Slovakia"                      "Slovenia"                     
#> [47] "Spain"                         "Svalbard and Jan Mayen"       
#> [49] "Sweden"                        "Switzerland"                  
#> [51] "Ukraine"                       "United Kingdom"               
#> 
#> $`North America`
#>  [1] "Anguilla"                         "Antigua and Barbuda"             
#>  [3] "Aruba"                            "Bahamas"                         
#>  [5] "Barbados"                         "Belize"                          
#>  [7] "Bermuda"                          "Bonaire, Sint Eustatius and Saba"
#>  [9] "Canada"                           "Cayman Islands"                  
#> [11] "Costa Rica"                       "Cuba"                            
#> [13] "Curaçao"                          "Dominica"                        
#> [15] "Dominican Republic"               "El Salvador"                     
#> [17] "Greenland"                        "Grenada"                         
#> [19] "Guadeloupe"                       "Guatemala"                       
#> [21] "Haiti"                            "Honduras"                        
#> [23] "Jamaica"                          "Martinique"                      
#> [25] "Mexico"                           "Montserrat"                      
#> [27] "Nicaragua"                        "Panama"                          
#> [29] "Puerto Rico"                      "Saint Barthélemy"                
#> [31] "Saint Kitts and Nevis"            "Saint Lucia"                     
#> [33] "Saint Martin (French part)"       "Saint Pierre and Miquelon"       
#> [35] "Saint Vincent and the Grenadines" "Trinidad and Tobago"             
#> [37] "Turks and Caicos Islands"         "United States"                   
#> [39] "Virgin Islands, British"          "Virgin Islands, U.S."            
#> 
#> $Oceania
#>  [1] "American Samoa"                      
#>  [2] "Australia"                           
#>  [3] "Clipperton Island"                   
#>  [4] "Cook Islands"                        
#>  [5] "Fiji"                                
#>  [6] "French Polynesia"                    
#>  [7] "Guam"                                
#>  [8] "Indian Ocean Territories"            
#>  [9] "Kiribati"                            
#> [10] "Marshall Islands"                    
#> [11] "Micronesia"                          
#> [12] "Nauru"                               
#> [13] "New Caledonia"                       
#> [14] "New Zealand"                         
#> [15] "Niue"                                
#> [16] "Norfolk Island"                      
#> [17] "Northern Mariana Islands"            
#> [18] "Palau"                               
#> [19] "Papua New Guinea"                    
#> [20] "Pitcairn Island"                     
#> [21] "Samoa"                               
#> [22] "Solomon Islands"                     
#> [23] "Tokelau"                             
#> [24] "Tonga"                               
#> [25] "Tuvalu"                              
#> [26] "United States Minor Outlying Islands"
#> [27] "Vanuatu"                             
#> [28] "Wallis and Futuna"                   
#> 
#> $`South America`
#>  [1] "Argentina"                                   
#>  [2] "Bolivia"                                     
#>  [3] "Brazil"                                      
#>  [4] "Chile"                                       
#>  [5] "Colombia"                                    
#>  [6] "Ecuador"                                     
#>  [7] "Falkland Islands (Malvinas)"                 
#>  [8] "French Guiana"                               
#>  [9] "Guyana"                                      
#> [10] "Paraguay"                                    
#> [11] "Peru"                                        
#> [12] "Saint Martin"                                
#> [13] "Sint Maarten (Dutch part)"                   
#> [14] "South Georgia and the South Sandwich Islands"
#> [15] "Suriname"                                    
#> [16] "Uruguay"                                     
#> [17] "Venezuela"                                   
#> 

accCountryNames('Europe')

#> $Europe
#>  [1] "Åland Islands"                 "Albania"                      
#>  [3] "Andorra"                       "Austria"                      
#>  [5] "Belarus"                       "Belgium"                      
#>  [7] "Bosnia and Herzegovina"        "Bulgaria"                     
#>  [9] "Croatia"                       "Czechia"                      
#> [11] "Denmark"                       "Estonia"                      
#> [13] "Faroe Islands"                 "Finland"                      
#> [15] "France"                        "Germany"                      
#> [17] "Gibraltar"                     "Greece"                       
#> [19] "Guernsey"                      "Holy See (Vatican City State)"
#> [21] "Hungary"                       "Iceland"                      
#> [23] "Ireland"                       "Isle of Man"                  
#> [25] "Italy"                         "Jersey"                       
#> [27] "Kosovo"                        "Latvia"                       
#> [29] "Liechtenstein"                 "Lithuania"                    
#> [31] "Luxembourg"                    "Macedonia"                    
#> [33] "Malta"                         "Moldova"                      
#> [35] "Monaco"                        "Montenegro"                   
#> [37] "Netherlands"                   "Norway"                       
#> [39] "Poland"                        "Portugal"                     
#> [41] "Romania"                       "Russia"                       
#> [43] "San Marino"                    "Serbia"                       
#> [45] "Slovakia"                      "Slovenia"                     
#> [47] "Spain"                         "Svalbard and Jan Mayen"       
#> [49] "Sweden"                        "Switzerland"                  
#> [51] "Ukraine"                       "United Kingdom"               
#>