Return the list of the continents and associated countries.

Return the list of the continents and associated countries.

accCountryNames(continent = NA, dbname = "gbif4crest_02")

Arguments

continent
A name of continent. Default is NA and returns a list of all the country names sorted by continent.

dbname
The name of the database. Default is 'gbif4crest_02'.

Value

A list where each element is a vector of corresponding country names.

Examples

accCountryNames()

#> [[1]]
#> [1] "Indian Ocean Territories" "Saint Helena"            
#> [3] "Sint Maarten"            
#> 
#> $Africa
#>  [1] "Algeria"                          "Angola"                          
#>  [3] "Benin"                            "Botswana"                        
#>  [5] "Burkina Faso"                     "Burundi"                         
#>  [7] "Cameroon"                         "Cape Verde"                      
#>  [9] "Central African Republic"         "Chad"                            
#> [11] "Comoros"                          "Democratic Republic of the Congo"
#> [13] "Djibouti"                         "Egypt"                           
#> [15] "Equatorial Guinea"                "Eritrea"                         
#> [17] "Eswatini"                         "Ethiopia"                        
#> [19] "Gabon"                            "Gambia"                          
#> [21] "Ghana"                            "Guinea"                          
#> [23] "Guinea Bissau"                    "Ivory Coast"                     
#> [25] "Kenya"                            "Lesotho"                         
#> [27] "Liberia"                          "Libya"                           
#> [29] "Madagascar"                       "Malawi"                          
#> [31] "Mali"                             "Mauritania"                      
#> [33] "Mauritius"                        "Morocco"                         
#> [35] "Mozambique"                       "Namibia"                         
#> [37] "Niger"                            "Nigeria"                         
#> [39] "Republic of the Congo"            "Rwanda"                          
#> [41] "Sao Tome and Principe"            "Senegal"                         
#> [43] "Seychelles"                       "Sierra Leone"                    
#> [45] "Somalia"                          "South Africa"                    
#> [47] "South Sudan"                      "Sudan"                           
#> [49] "Tanzania"                         "Togo"                            
#> [51] "Tunisia"                          "Uganda"                          
#> [53] "Western Sahara"                   "Zambia"                          
#> [55] "Zimbabwe"                        
#> 
#> $Antarctica
#> [1] "Antarctica"                              
#> [2] "French Southern and Antarctic Lands"     
#> [3] "Heard Island and McDonald Islands"       
#> [4] "South Georgia and South Sandwich Islands"
#> 
#> $Asia
#>  [1] "Afghanistan"                    "Bahrain"                       
#>  [3] "Bangladesh"                     "Bhutan"                        
#>  [5] "British Indian Ocean Territory" "Brunei"                        
#>  [7] "Cambodia"                       "China"                         
#>  [9] "East Timor"                     "Hong Kong"                     
#> [11] "India"                          "Indonesia"                     
#> [13] "Iran"                           "Iraq"                          
#> [15] "Israel"                         "Japan"                         
#> [17] "Jordan"                         "Kazakhstan"                    
#> [19] "Kuwait"                         "Kyrgyzstan"                    
#> [21] "Laos"                           "Lebanon"                       
#> [23] "Malaysia"                       "Maldives"                      
#> [25] "Mongolia"                       "Myanmar"                       
#> [27] "Nepal"                          "North Korea"                   
#> [29] "Oman"                           "Pakistan"                      
#> [31] "Palestinian Territory"          "Philippines"                   
#> [33] "Qatar"                          "Saudi Arabia"                  
#> [35] "Singapore"                      "South Korea"                   
#> [37] "Sri Lanka"                      "Syria"                         
#> [39] "Taiwan"                         "Tajikistan"                    
#> [41] "Thailand"                       "Turkmenistan"                  
#> [43] "United Arab Emirates"           "Uzbekistan"                    
#> [45] "Vietnam"                        "Yemen"                         
#> 
#> $Europe
#>  [1] "Aland Islands"          "Albania"                "Armenia"               
#>  [4] "Austria"                "Azerbaijan"             "Belarus"               
#>  [7] "Belgium"                "Bosnia and Herzegovina" "Bulgaria"              
#> [10] "Croatia"                "Cyprus"                 "Czech Republic"        
#> [13] "Denmark"                "Estonia"                "Faroe Islands"         
#> [16] "Finland"                "France"                 "Georgia"               
#> [19] "Germany"                "Greece"                 "Hungary"               
#> [22] "Iceland"                "Ireland"                "Isle of Man"           
#> [25] "Italy"                  "Kazakhstan"             "Kosovo"                
#> [28] "Latvia"                 "Lithuania"              "Luxembourg"            
#> [31] "Macedonia"              "Malta"                  "Moldova"               
#> [34] "Montenegro"             "Netherlands"            "Norway"                
#> [37] "Poland"                 "Portugal"               "Romania"               
#> [40] "Russia"                 "Serbia"                 "Slovakia"              
#> [43] "Slovenia"               "Spain"                  "Sweden"                
#> [46] "Switzerland"            "Turkey"                 "Ukraine"               
#> [49] "United Kingdom"        
#> 
#> $`North America`
#>  [1] "Anguilla"                         "Antigua and Barbuda"             
#>  [3] "Barbados"                         "Belize"                          
#>  [5] "Bermuda"                          "Canada"                          
#>  [7] "Cayman Islands"                   "Commonwealth of the Bahamas"     
#>  [9] "Costa Rica"                       "Cuba"                            
#> [11] "Curaçao"                          "Dominica"                        
#> [13] "Dominican Republic"               "El Salvador"                     
#> [15] "Greenland"                        "Grenada"                         
#> [17] "Guatemala"                        "Haiti"                           
#> [19] "Honduras"                         "Jamaica"                         
#> [21] "Mexico"                           "Nicaragua"                       
#> [23] "Panama"                           "Puerto Rico"                     
#> [25] "Saint Kitts and Nevis"            "Saint Lucia"                     
#> [27] "Saint Pierre and Miquelon"        "Saint Vincent and the Grenadines"
#> [29] "Trinidad and Tobago"              "Turks and Caicos Islands"        
#> [31] "United States of America"         "United States Virgin Islands"    
#> 
#> $Oceania
#>  [1] "American Samoa"                 "Australia"                     
#>  [3] "Cook Islands"                   "Federated States of Micronesia"
#>  [5] "Fiji"                           "French Polynesia"              
#>  [7] "Guam"                           "Kiribati"                      
#>  [9] "Marshall Islands"               "New Caledonia"                 
#> [11] "New Zealand"                    "Niue"                          
#> [13] "Northern Mariana Islands"       "Palau"                         
#> [15] "Papua New Guinea"               "Pitcairn Islands"              
#> [17] "Samoa"                          "Solomon Islands"               
#> [19] "Tonga"                          "Vanuatu"                       
#> [21] "Wallis and Futuna"             
#> 
#> $`South America`
#>  [1] "Argentina"        "Bolivia"          "Brazil"           "Chile"           
#>  [5] "Colombia"         "Ecuador"          "Falkland Islands" "Guyana"          
#>  [9] "Paraguay"         "Peru"             "Suriname"         "Uruguay"         
#> [13] "Venezuela"       
#> 

accCountryNames('Europe')

#> $Europe
#>  [1] "Aland Islands"          "Albania"                "Armenia"               
#>  [4] "Austria"                "Azerbaijan"             "Belarus"               
#>  [7] "Belgium"                "Bosnia and Herzegovina" "Bulgaria"              
#> [10] "Croatia"                "Cyprus"                 "Czech Republic"        
#> [13] "Denmark"                "Estonia"                "Faroe Islands"         
#> [16] "Finland"                "France"                 "Georgia"               
#> [19] "Germany"                "Greece"                 "Hungary"               
#> [22] "Iceland"                "Ireland"                "Isle of Man"           
#> [25] "Italy"                  "Kazakhstan"             "Kosovo"                
#> [28] "Latvia"                 "Lithuania"              "Luxembourg"            
#> [31] "Macedonia"              "Malta"                  "Moldova"               
#> [34] "Montenegro"             "Netherlands"            "Norway"                
#> [37] "Poland"                 "Portugal"               "Romania"               
#> [40] "Russia"                 "Serbia"                 "Slovakia"              
#> [43] "Slovenia"               "Spain"                  "Sweden"                
#> [46] "Switzerland"            "Turkey"                 "Ukraine"               
#> [49] "United Kingdom"        
#>