Extract distributions from the database

This function will extract the distributions of all the species composing each taxon and return them as a list.

crest.get_modern_data(
  pse,
  taxaType,
  climate,
  df = NA,
  ai.sqrt = FALSE,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  continents = NA,
  countries = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  minGridCells = 20,
  selectedTaxa = NA,
  site_info = c(NA, NA),
  site_name = NA,
  dbname = "gbif4crest_02",
  verbose = TRUE
)

Arguments

pse
A pollen-Species equivalency table. See createPSE for details.

taxaType
A numerical index (between 1 and 6) to define the type of palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids, 4 for foraminifers, 5 for diatoms and 6 for rodents. The example dataset uses taxaType=0 (pseudo-data). Default is 1.

climate
A vector of the climate variables to extract. See accClimateVariables for the list of accepted values.

df
A data frame containing the data to reconstruct (counts, percentages or presence/absence data).

ai.sqrt
A boolean to indicate whether ai values should be square-root transformed (default FALSE).

xmn
The coordinates defining the study area.

xmx
The coordinates defining the study area.

ymn
The coordinates defining the study area.

ymx
The coordinates defining the study area.

continents
A vector of the continent names defining the study area.

countries
A vector of the country names defining the study area.

realms
A vector of the studied botanical realms defining the study area.

biomes
A vector of the studied botanical biomes defining the study area.

ecoregions
A vector of the studied botanical ecoregions defining the study area.

minGridCells
The minimum number of unique presence data necessary to estimate a species' climate response. Default is 20.

selectedTaxa
A data frame assigns which taxa should be used for each variable (1 if the taxon should be used, 0 otherwise). The colnames should be the climate variables' names and the rownames the taxa names. Default is 1 for all taxa and all variables.

site_info
A vector containing the coordinates of the study site. Default c(NA, NA).

site_name
The name of the dataset (default NA).

dbname
The name of the database. Default is 'gbif4crest_02'.

verbose
A boolean to print non-essential comments on the terminal (default TRUE).

Value

A crestObj object containing the spatial distributions.

Examples

data(crest_ex_pse)
data(crest_ex_selection)
data(crest_ex)
x <- crest.get_modern_data( df = crest_ex,
  pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example",
  verbose = FALSE
)
x

#> $inputs
#> [1] "df"           "taxa.name"    "x"            "pse"          "selectedTaxa"
#> [6] "x.name"      
#> 
#> $parameters
#>  [1] "climate"               "taxaType"              "xmn"                  
#>  [4] "xmx"                   "ymn"                   "ymx"                  
#>  [7] "continents"            "countries"             "realms"               
#> [10] "biomes"                "ecoregions"            "taxWeight"            
#> [13] "minGridCells"          "bin_width"             "shape"                
#> [16] "npoints"               "geoWeighting"          "climateSpaceWeighting"
#> [19] "presenceThreshold"     "uncertainties"        
#> 
#> $modelling
#> [1] "taxonID2proxy" "climate_space" "pdfs"          "weights"      
#> [5] "xrange"        "distributions"
#> 
#> $reconstructions
#> NULL
#> 
#> $misc
#> [1] "dbname"     "taxa_notes" "site_info" 
#> 

lapply(x$modelling$distributions, head)

#> $Taxon1
#>   taxonid longitude latitude bio1 bio12
#> 1       1      0.25     8.75 16.0   477
#> 2       1      0.25    10.25 17.7   594
#> 3       1      0.75     8.75 17.3   591
#> 4       1      1.25    11.75 18.8   814
#> 5       1      1.25    12.25 15.8   768
#> 6       1      2.75     7.75 18.6   500
#> 
#> $Taxon2
#>   taxonid longitude latitude bio1 bio12
#> 1       2      5.25     3.75 22.4   240
#> 2       2      7.25     0.25 21.0     5
#> 3       2      8.25    10.75 22.1   575
#> 4       2      8.75     4.25 22.3   302
#> 5       2      9.25     3.25 21.3    98
#> 6       2      9.25     6.25 21.6   394
#> 
#> $Taxon3
#>   taxonid longitude latitude bio1 bio12
#> 1       3      0.25     0.25 15.4     3
#> 2       3      0.25     2.75 13.4   325
#> 3       3      0.25     3.25 12.3   329
#> 4       3      0.25     3.75 15.6   242
#> 5       3      0.25     4.75 13.7   371
#> 6       3      0.25     8.25 11.5   538
#> 
#> $Taxon4
#>   taxonid longitude latitude bio1 bio12
#> 1       4      0.25     7.25 14.7   412
#> 2       4      0.25    10.75 14.7   750
#> 3       4      0.25    11.25 14.8   655
#> 4       4      0.25    12.25 17.0   729
#> 5       4      0.25    13.75 16.2   981
#> 6       4      0.25    14.25 14.0   910
#> 
#> $Taxon5
#>   taxonid longitude latitude bio1 bio12
#> 1       5      0.25     0.75 15.1    38
#> 2       5      0.25     1.25 18.2    37
#> 3       5      0.25     1.75 13.6   137
#> 4       5      0.25     2.25 15.5   178
#> 5       5      0.25     3.25 12.3   329
#> 6       5      0.25     3.75 15.6   242
#> 
#> $Taxon6
#>   taxonid longitude latitude bio1 bio12
#> 1       6      0.25     0.25 15.4     3
#> 2       6      0.25     0.75 15.1    38
#> 3       6      0.25     1.25 18.2    37
#> 4       6      0.25     2.25 15.5   178
#> 5       6      0.25     3.25 12.3   329
#> 6       6      0.25     3.75 15.6   242
#>