Extract taxonID(s) corresponding to the taxonomic description

Extract all possible TaxonIDs corresponding to the provided taxonomical description, which can be at the family, the genus or the species levels.

getDistribTaxa(
  taxIDs,
  climate = NA,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  continents = NA,
  countries = NA,
  basins = NA,
  sectors = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  elev_min = NA,
  elev_max = NA,
  elev_range = NA,
  year_min = 1900,
  year_max = 2021,
  nodate = TRUE,
  type_of_obs = c(1, 2, 7, 8),
  dbname = "gbif4crest_02"
)

Arguments

taxIDs

A vector of accepted Taxa IDs (as returned by getTaxonID).

climate

A vector of the climate variables to extract. See accClimateVariables for the list of accepted values.

xmn, xmx, ymn, ymx

The coordinates defining the study area.

continents

A vector of the continent names defining the study area.

countries

A vector of the country names defining the study area.

basins

A vector of the ocean names defining the study area.

sectors

A vector of the marine sector names defining the study area.

realms

A vector of the studied botanical realms defining the study area.

biomes

A vector of the studied botanical biomes defining the study area.

ecoregions

A vector of the studied botanical ecoregions defining the study area.

elev_min, elev_max

Parameters to only selected grid cells with an elevation higher than elev_min or lower than elev_max (default is 'NA ).

elev_range

Parameters discard the grid cell with a high elevation range (default is NA).

year_min, year_max

The oldest and youngest occurrences accepted (default is 1900-2021).

nodate

A boolean to accept occurrences without a date (can overlap with occurrences with a date; default TRUE).

type_of_obs

The type of observation to use in the study. 1: human observation, 2: observation, 3: preserved specimen, 4: living specimen, 5: fossil specimen, 6: material sample, 7: machine observation, 8: literature, 9: unknown (Default c(1, 2, 3, 8, 9))

dbname

The name of the data source database.

Value

A matrix of occurrence records with the associated climate.

See also

Examples

if (FALSE) {
  taxIDs <- getTaxonID("Zamiaceae", "Ceratozamia")
  distrib <- getDistribTaxa(taxIDs, "bio1", -90, 90, -90, 90,
    continents = "Europe",
    countries = c("Germany", "Netherlands", "Sweden"),
    realms = "Palaearctic"
  )
  distrib
}