Extract taxonID(s) corresponding to the taxonomic description

Extract all possible TaxonIDs corresponding to the provided taxonomical description, which can be at the family, the genus or the species levels.

getTaxonID(
  family = "",
  genus = "",
  species = "",
  taxaType = 1,
  dbname = "gbif4crest_02"
)

Arguments

family
The name of the family.

genus
The name of the genus.

species
The name of the species.

taxaType
A numerical index (between 1 and 6) to define the type of palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids, 4 for foraminifers, 5 for diatoms and 6 for rodents. The example dataset uses taxaType=0 (pseudo-data). Default is 1.

dbname
The name of the data source database.

Value

A vector of unique taxonIDs.

Examples

if (FALSE) {
  getTaxonID("Zamiaceae")
  getTaxonID("Zamiaceae", "Ceratozamia")
  ## \code{taxaType = 2} searches for beetles and not plants, so the next line returns nothing.
  getTaxonID("Zamiaceae", "Ceratozamia", taxaType = 2)
}